unsigned comparisons i guesss

+++ lab-shell-pipeline/tsh.c
@@ -135,14 +135,14 @@ void eval(char *cmdline)
         if(pid1==0){
             // Check the command for any input or output redirection, and perform that redirection.
             FILE * fp;
-            if(stdin_redir[0] > 0){
+            if((int)stdin_redir[0] > 0){
                 // redirect stdin to stdin_redir[i]
 
                 fp = fopen(argv[stdin_redir[0]],"r");
                 dup2(fileno(fp),STDIN_FILENO); // STDIN_FILENO is 0
                 // dup2(stdin_redir[0],1); // TODO: i ------------------------------------------
             }
-            if (stdout_redir[0] > 0){
+            if ((int)stdout_redir[0] > 0){
                 // redirect stdout to stddout_redir[i]
                 
                 fp = fopen(argv[stdout_redir[0]],"w");
@@ -188,7 +188,7 @@ void eval(char *cmdline)
     if(pid1==0){
         // Check the command for any input or output redirection, and perform that redirection.
         FILE * fp;
-        if(stdin_redir[0] > 0){
+        if((int)stdin_redir[0] > 0){
             // redirect stdin to stdin_redir[i]
 
             fp = fopen(argv[stdin_redir[0]],"r");
@@ -234,7 +234,7 @@ void eval(char *cmdline)
             //     dup2(fileno(fp),STDIN_FILENO); // STDIN_FILENO is 0
             //     // dup2(stdin_redir[0],1); // TODO: i 
             // }
-            if (stdout_redir[1] > 0){
+            if ((int)stdout_redir[1] > 0){
                 // redirect stdout to stddout_redir[i]
                 fprintf("%d\n", stdout_redir[1]);
